requirements.txt
# ─── requirements.txt ───

pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0

# (We install PyQt5 via apt, so we do NOT list pyqt5 here)
# pyqt5>=5.15.0

# Optional extras:
# scipy>=1.7.0
# pandas>=1.3.0
# opencv-python>=4.5.0.68
